Moved ip to classes since it fits in more there
